backend/requirements.txt
uvicorn
fastapi
scapy
python-nmap
pydantic
confluent-kafka
requests